requirements.txt
requests
rich
psutil
py_cpuinfo
GPUtil==1.4.0
